+++ src/main/java/org/truffleruby/parser/BodyTranslator.java
@@ -721,7 +721,7 @@ public class BodyTranslator extends Translator {
         final SourceIndexLength enclosingSourceSection = enclosing(sourceSection, children.toArray(RubyNode.EMPTY_ARRAY));
 
         RubyCallNodeParameters callParameters = new RubyCallNodeParameters(receiver, methodName, argumentsAndBlock.getBlock(), argumentsAndBlock.getArguments(), argumentsAndBlock.isSplatted(), privately || ignoreVisibility, isVCall, node.isLazy(), isAttrAssign);
-        RubyNode translated = Translator.withSourceSection(enclosingSourceSection, context.getCoreMethods().createCallNode(callParameters));
+        RubyNode translated = Translator.withSourceSection(enclosingSourceSection, context.getCoreMethods().createCallNode(callParameters, environment));
 
         translated = wrapCallWithLiteralBlock(argumentsAndBlock, translated);
 
@@ -2378,7 +2378,7 @@ public class BodyTranslator extends Translator {
             default: {
                 final SourceIndexLength sourceSection = node.getPosition();
                 final RubyCallNodeParameters callParameters = new RubyCallNodeParameters(lhs, node.getOperator(), null, new RubyNode[] { rhs }, false, true);
-                final RubyNode opNode = context.getCoreMethods().createCallNode(callParameters);
+                final RubyNode opNode = context.getCoreMethods().createCallNode(callParameters, environment);
                 final RubyNode ret = ((ReadConstantNode) lhs).makeWriteNode(opNode);
                 ret.unsafeSetSourceSection(sourceSection);
                 return addNewlineIfNeeded(node, ret);
@@ -3055,7 +3055,8 @@ public class BodyTranslator extends Translator {
             argumentsTranslated[i] = arguments[i].accept(this);
         }
 
-        final RubyNode ret = new YieldExpressionNode(unsplat, argumentsTranslated);
+        RubyNode readBlock = environment.findLocalVarOrNilNode(TranslatorEnvironment.METHOD_BLOCK_NAME, node.getPosition());
+        final RubyNode ret = new YieldExpressionNode(unsplat, argumentsTranslated, readBlock);
         ret.unsafeSetSourceSection(node.getPosition());
         return addNewlineIfNeeded(node, ret);
     }