+++ src/main/java/org/truffleruby/parser/LoadArgumentsTranslator.java
@@ -86,6 +86,7 @@ public class LoadArgumentsTranslator extends Translator {
     }
 
     private final boolean isProc;
+    private final boolean isMethod;
     private final BodyTranslator methodBodyTranslator;
     private final Deque<ArraySlot> arraySlotStack = new ArrayDeque<>();
 
@@ -104,9 +105,10 @@ public class LoadArgumentsTranslator extends Translator {
     private State state;
     private boolean firstOpt = false;
 
-    public LoadArgumentsTranslator(Node currentNode, ArgsParseNode argsNode, RubyContext context, Source source, ParserContext parserContext, boolean isProc, BodyTranslator methodBodyTranslator) {
+    public LoadArgumentsTranslator(Node currentNode, ArgsParseNode argsNode, RubyContext context, Source source, ParserContext parserContext, boolean isProc, boolean isMethod, BodyTranslator methodBodyTranslator) {
         super(currentNode, context, source, parserContext);
         this.isProc = isProc;
+        this.isMethod = isMethod;
         this.methodBodyTranslator = methodBodyTranslator;
         this.argsNode = argsNode;
         this.required = argsNode.getRequiredCount();
@@ -153,6 +155,12 @@ public class LoadArgumentsTranslator extends Translator {
                 sequence.add(args[i].accept(this));
                 index++;
             }
+        }
+
+        // Do this before handling optional arguments as one might get
+        // its default value via a `yield`.
+        if (isMethod) {
+            sequence.add(saveMethodBlockArg());
         }
 
         final int optArgCount = argsNode.getOptionalArgsCount();
@@ -331,6 +339,12 @@ public class LoadArgumentsTranslator extends Translator {
         }
 
         final FrameSlot slot = methodBodyTranslator.getEnvironment().getFrameDescriptor().findFrameSlot(node.getName());
+        return new WriteLocalVariableNode(slot, readNode);
+    }
+
+    public RubyNode saveMethodBlockArg() {
+        final RubyNode readNode = new ReadBlockNode(context.getCoreLibrary().getNil());
+        final FrameSlot slot = methodBodyTranslator.getEnvironment().getFrameDescriptor().findOrAddFrameSlot(TranslatorEnvironment.METHOD_BLOCK_NAME);
         return new WriteLocalVariableNode(slot, readNode);
     }